+++ rich/measure.py
@@ -57,6 +57,8 @@ class Measurement(NamedTuple):
 
         if isinstance(renderable, str):
             renderable = console.render_str(renderable)
+
+        renderable = getattr(renderable, "__rich__", renderable)
         if is_renderable(renderable):
             get_console_width = getattr(renderable, "__measure__", None)
             if get_console_width is not None: